requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.2
colorama==0.4.6
Faker==37.5.3
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-MySQLdb==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
tzdata==2025.2
Werkzeug==3.1.3
